Refactor UnifiedGlamourManager structure and formatting
